@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 
 ruby "2.1.0"
 
-gem 'rails', '4.0.3'
+gem 'rails', '~> 4.1.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -13,10 +13,12 @@ group :development, :test do
   gem "bullet"
 end
 
+
 group :test do
-  gem "factory_girl_rails", "~> 4.4.0"
+  gem "factory_girl_rails", "~> 4.4.1"
   gem 'cucumber-rails', "1.4.0", :require => false
   gem 'database_cleaner', "1.2.0"
+  gem 'minitest'
   gem 'shoulda-matchers'
 end
 
@@ -29,9 +31,9 @@ gem 'will_paginate', '~> 3.0'
 
 gem "thin"
 
-gem 'sass-rails',   '~> 4.0.0'
+gem 'sass-rails',   '~> 4.0.0', '>= 4.0.2'
 gem 'coffee-rails', '~> 4.0.0'
-gem 'bootstrap-sass', '~> 3.1.1'
+gem 'bootstrap-sass', '~> 3.1.1.0'
 
 gem 'uglifier', '>= 1.3.0'
 
