@@ -5,7 +5,7 @@ require 'winery.rb'
 class WineScoreExporter
 
   def create_wine_sentence(wine, winery)
-    s = [winery.name.gsub('.', ''), wine.grapes.join(', '), wine.year]
+    s = [winery.name.gsub('.', ''), wine.grapes_to_s(', '), wine.year]
     s << wine.other.gsub('.', '') unless wine.other.nil? or wine.other.empty?
     s << wine.region unless wine.region.nil? or wine.region.empty?
     s << "LCBO# " + wine.lcbo_code unless wine.lcbo_code.empty?
@@ -39,7 +39,7 @@ class WineScoreExporter
       l << "#{country}"
       l << NEW_LINE
 
-      # load all wines for country country
+      # load all wines for country
       wines = Wine.where("country = ?", country).order('winery_id asc')
 
       wines.each do |wine|
