@@ -14,18 +14,18 @@ describe ReportsController do
     end
 
     it "assigns all wines in cellar to @wines variable by default" do
-      s1 = create(:bottle_drank)
-      s2 = create(:bottle_in_cellar_later)
-      s3 = create(:bottle_in_cellar_sooner)
+      s1 = create(:alvento_sauvignonblanc_bottle_drank)
+      s2 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s3 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
 
       get 'cellar'
       assigns[:wines].should == [s3.wine, s2.wine]
     end
 
     it "assigns wines ready to drink soon when told to" do
-      s1 = create(:bottle_drank)
-      s2 = create(:bottle_in_cellar_later)
-      s3 = create(:bottle_in_cellar_sooner)
+      s1 = create(:alvento_sauvignonblanc_bottle_drank)
+      s2 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s3 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
 
       get 'cellar', :term => 'soon'
       assigns[:wines].should == [s3.wine]
@@ -34,9 +34,9 @@ describe ReportsController do
 
   describe "GET 'maturity_profile'" do
     it "generates a breakdown of bottles vs. year" do
-      s1 = create(:bottle_in_cellar_later)
-      s2 = create(:bottle_in_cellar_sooner)
-      s3 = create(:bottle_in_cellar_now)
+      s1 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s2 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
+      s3 = create(:imocali_vernaccia_bottle_now)
 
       get 'maturity_profile'
 
@@ -49,9 +49,9 @@ describe ReportsController do
 
   describe "GET 'score_breakdown'" do
     it "assigns score breakdown to param" do
-      s1 = create(:bottle_in_cellar_later)
-      s2 = create(:bottle_in_cellar_sooner)
-      s3 = create(:bottle_in_cellar_now)
+      s1 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s2 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
+      s3 = create(:imocali_vernaccia_bottle_now)
 
       get 'score_breakdown'
 
@@ -74,9 +74,9 @@ describe ReportsController do
     end
 
     it "assigns country report data to @country_summaries" do
-      create(:bottle_drank)
-      create(:bottle_drank_2)
-      create(:bottle_drank_3)
+      create(:alvento_sauvignonblanc_bottle_drank)
+      create(:kenwood_merlot_drank)
+      create(:vineland_sauvignonblanc_drank)
 
       get 'country'
 
@@ -96,7 +96,7 @@ describe ReportsController do
     end
 
     it "assigns country report data to @wines" do
-      create(:bottle_drank)
+      create(:alvento_sauvignonblanc_bottle_drank)
 
       get 'this_day_in_wine'
 
@@ -116,7 +116,7 @@ describe ReportsController do
     end
 
     it "assigns country report data to @wines" do
-      create(:bottle_drank_2)
+      create(:kenwood_merlot_drank)
 
       get 'favourite_wines'
 
