@@ -14,21 +14,28 @@ describe Wine do
     end
   end
 
+  describe "grapes" do
+    it "must be able to output its list of grapes as a string" do
+      w = create(:kenwood_merlot)
+      expect(w.grapes_to_s).to include('Merlot')
+    end
+  end
+
   describe "bottles" do
     it "knows its favourite bottles don't have scores less than 90" do
-      b1 = create(:bottle_drank)
+      b1 = create(:alvento_sauvignonblanc_bottle_drank)
 
       expect(b1.wine.favourite_bottles).to eq([])
     end
 
     it "knows its favourite bottles have scores greater or equal to 90" do
-      b1 = create(:bottle_drank_2)
+      b1 = create(:kenwood_merlot_drank)
 
       expect(b1.wine.favourite_bottles).to eq([b1])
     end
 
     it "knows bottles that have been drunk" do
-      b1 = create(:bottle_drank)
+      b1 = create(:alvento_sauvignonblanc_bottle_drank)
       wine = b1.wine
       b2 = Bottle.new(reviewdate: 2.days.from_now.strftime('%d %b %Y'))
       wine.bottles << b2
@@ -38,51 +45,51 @@ describe Wine do
 
   describe "finders" do
     it "finds all wines with scores greater than 90" do
-      b1 = create(:bottle_drank)
-      b2 = create(:bottle_drank_2)
+      b1 = create(:alvento_sauvignonblanc_bottle_drank)
+      b2 = create(:kenwood_merlot_drank)
 
       expect(Wine.find_favourites 90).to eq([b2.wine])
     end
 
     it "finds all wines with bottles indicating they are in the cellar" do
-      s1 = create(:bottle_drank)
-      s2 = create(:bottle_in_cellar_later)
-      s3 = create(:bottle_in_cellar_sooner)
-      s4 = create(:bottle_in_cellar_now)
+      s1 = create(:alvento_sauvignonblanc_bottle_drank)
+      s2 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s3 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
+      s4 = create(:imocali_vernaccia_bottle_now)
 
       expect(Wine.find_in_cellar).to eq([s3.wine, s4.wine, s2.wine])
     end
 
     it "finds all wines that will be ready to drink in the next two years" do
-      s1 = create(:bottle_in_cellar_later)
-      s2 = create(:bottle_in_cellar_sooner)
-      s3 = create(:bottle_in_cellar_now)
+      s1 = create(:vineland_sauvignonblanc_bottle_in_cellar_later)
+      s2 = create(:cavesprings_sauvignonblanc_bottle_in_cellar_sooner)
+      s3 = create(:imocali_vernaccia_bottle_now)
 
       expect(Wine.find_in_cellar_ready_to_drink).to eq([s2.wine, s3.wine])      
     end
 
     it "searches all wines by winery name" do
-      wine = create(:wine_with_grapes_a)
+      wine = create(:alvento_sauvignonblanc)
 
       expect(Wine.search_for_wine('lvento', '', '')[0]).to eq(wine)
     end
 
     it "searches all wines by wine other info" do
-      wine = create(:wine_with_grapes_a)
+      wine = create(:alvento_sauvignonblanc)
 
       expect(Wine.search_for_wine('Reserve', '', '')[0]).to eq(wine)
     end
 
     it "searches all wines by region" do
-      wine = create(:wine_with_grapes_a)
+      wine = create(:alvento_sauvignonblanc)
 
       expect(Wine.search_for_wine('VQA Niagara', '', '')[0]).to eq(wine)
     end
 
     it "searches only for wines with reviews after a given date" do
-      wine = create(:bottle_drank).wine
-      create(:bottle_drank_2)
-      create(:bottle_drank_3)
+      wine = create(:alvento_sauvignonblanc_bottle_drank).wine
+      create(:kenwood_merlot_drank)
+      create(:vineland_sauvignonblanc_drank)
       review_from = -1.days.from_now.strftime('%d %b %Y')
 
       matching_wines = Wine.search_for_wine('', review_from, '')
@@ -91,9 +98,9 @@ describe Wine do
     end
 
     it "searches only for wines with reviews before a given date" do
-      wine1 = create(:bottle_drank).wine
-      create(:bottle_drank_2)
-      wine3 = create(:bottle_drank_3).wine
+      wine1 = create(:alvento_sauvignonblanc_bottle_drank).wine
+      create(:kenwood_merlot_drank)
+      wine3 = create(:vineland_sauvignonblanc_drank).wine
       review_to = -3.days.from_now.strftime('%d %b %Y')
 
       matching_wines = Wine.search_for_wine('', '', review_to)
@@ -107,9 +114,9 @@ describe Wine do
     end
 
     it "searches only for wines within reviews in a range of dates" do
-      wine1 = create(:bottle_drank).wine
-      wine2 = create(:bottle_drank_2).wine
-      create(:bottle_drank_3)
+      wine1 = create(:alvento_sauvignonblanc_bottle_drank).wine
+      wine2 = create(:kenwood_merlot_drank).wine
+      create(:vineland_sauvignonblanc_drank)
       review_from = -3.days.from_now.strftime('%d %b %Y')
       review_to = -1.days.from_now.strftime('%d %b %Y')
 
@@ -119,21 +126,21 @@ describe Wine do
     end
 
     it "finds wines matching what is parsed from file" do
-      w = create(:wine_with_grapes_a)
+      w = create(:alvento_sauvignonblanc)
       winery = Winery.find_by_name('Alvento')
       
       expect(Wine.find_by_winery_year_lcbo_code(winery.id, 2011, '123456')).to eq([w])
     end
 
     it "doesn't find results when what is parsed from file doesn't have lcbo code" do
-      w = create(:wine_with_grapes_b)
+      w = create(:vineland_sauvignonblanc)
       winery = Winery.find_by_name('Vineland Estates')
       
       expect(Wine.find_by_winery_year_lcbo_code(winery.id, 2011, '')).to eq([])
     end
-    
+
     it "finds all bottles drunk on today's date" do
-      b = create(:bottle_drank)
+      b = create(:alvento_sauvignonblanc_bottle_drank)
 
       expect(Wine.find_drank_this_day[0]).to eq(b.wine)
     end
@@ -141,8 +148,8 @@ describe Wine do
 
   describe "filters" do
     it "filters by grape type" do
-      create(:wine_with_grapes_a)
-      w = create(:wine_with_grapes_d)
+      create(:alvento_sauvignonblanc)
+      w = create(:imocali_vernaccia)
 
       found = Wine.filter_by_grapes_country('Vernaccia', '')
       expect(found[0]).to eq(w)
@@ -150,8 +157,8 @@ describe Wine do
     end
 
     it "filters by country" do
-      create(:wine_with_grapes_a)
-      w = create(:wine_with_grapes_d)
+      create(:alvento_sauvignonblanc)
+      w = create(:imocali_vernaccia)
 
       found = Wine.filter_by_grapes_country('', 'Italy')
       expect(found[0]).to eq(w)
@@ -159,9 +166,10 @@ describe Wine do
     end
 
     it "filters by grape type and country" do
-      create(:wine_with_grapes_c)
-      create(:wine_with_grapes_e)
-      w = create(:wine_lailey_merlot)
+      create(:vineland_sauvignonblanc)
+      # create(:kenwood_merlot)
+      create(:imocali_vernaccia)
+      w = create(:lailey_merlot)
 
       # FIXME figure out why nothing gets returned
       found = Wine.filter_by_grapes_country('Merlot', 'Canada')
